A really really bad LSP proxy?
It's become very apparent that this isn't the Right Way, so I'm going to stop this line of exploration. Intercepting LSP messages, though possibly a better option (separate from RLS, can work with any LSP Server, etc.) is going to involve duplicating a fair bit of the logic in RLS and likely will always be another non-negligible failure point.
Instead, it's probably best to at least try to get a wsl option upstreamed into RLS or, failing that, modes added to the LSP client(s).

This repo will probably eventually morph into something that generates and installs shims for cargo and rustup in the right places (for IntelliJ and co.)